+++ worker/solver_worker.ml
@@ -10,7 +10,7 @@ module Solver_request = struct
   module Worker = Solver_service_api.Worker
   module Log = Solver_service_api.Solver.Log
   module Service = Solver_service.Service.Make (Solver_service.Opam_repository)
-  module Worker_process = Solver_service.Internal_worker.Worker_process
+  module Solver_process = Solver_service.Internal_worker.Solver_process
 
   type t =
     ( Service.Epoch.t,
@@ -27,7 +27,7 @@ module Solver_request = struct
             Solver_service.Remote_commit.list_to_string commits;
           |] )
       in
-      Worker_process.create cmd
+      Solver_process.create cmd
     in
     let create commits =
       Service.Epoch.create ~n_workers ~create_worker commits